Replaced deprecated methods for iOS 7.0 and up
Minimum iOS version support is iOS 7.0.

Methods deprecated were replaced while preserving the existing methods.
This required changes in the SPAudioEngine and SPTextField. I changed
to the audio engine to use a singleton to hold values.
NSAttributedString was used to replaced the old methods. To make it
easier, I added new member variables for paragraph style and font.